Use corestore for storage.
Keep the public API of multifeed but internally use corestore for
storage. For storing feed names and which feeds are part of a multifeed,
support external handlers to store and load that information. Have a
default handler that uses a hypercore for this. The hypercore is
identified as the default feed of a corestore namespace for the root
key.